Reorganize classnames and files
